:art: create a SearchResultInterface that the children fragments implement.
move the search handling up to the parent SearchFragment, and just have it call the child fragment's search methods.
