+++ app/src/main/java/com/codepath/instagram/fragments/SearchFragment.java
@@ -5,13 +5,29 @@ import android.os.Bundle;
 import android.support.design.widget.TabLayout;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
+import android.support.v4.view.MenuItemCompat;
 import android.support.v4.view.ViewPager;
+import android.support.v7.widget.SearchView;
+import android.text.TextUtils;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
 import com.codepath.instagram.R;
+import com.codepath.instagram.core.MainApplication;
 import com.codepath.instagram.helpers.SmartFragmentStatePagerAdapter;
+import com.codepath.instagram.helpers.Utils;
+import com.codepath.instagram.models.InstagramSearchTag;
+import com.loopj.android.http.JsonHttpResponseHandler;
+
+import org.apache.http.Header;
+import org.json.JSONObject;
+
+import java.util.List;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
@@ -81,4 +97,35 @@ public class SearchFragment extends Fragment {
         return view;
     }
 
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        setHasOptionsMenu(true);
+    }
+
+    @Override
+    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+        inflater.inflate(R.menu.menu_search, menu);
+
+        MenuItem searchItem = menu.findItem(R.id.action_search);
+        SearchView searchView = (SearchView) MenuItemCompat.getActionView(searchItem);
+
+        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+            @Override
+            public boolean onQueryTextSubmit(String query) {
+                // perform query here
+                return false;
+            }
+
+            @Override
+            public boolean onQueryTextChange(String newText) {
+                return ((SearchResultInterface) mSearchPagerAdapter.getRegisteredFragment(
+                        mSearchTL.getSelectedTabPosition()
+                )).onSearchQuery(newText);
+            }
+        });
+
+        super.onCreateOptionsMenu(menu, inflater);
+    }
 }
