+++ app/src/main/java/com/codepath/instagram/fragments/SearchResultInterface.java
@@ -1,0 +1,8 @@
+package com.codepath.instagram.fragments;
+
+/**
+ * Created by koulmomo on 12/2/15.
+ */
+public interface SearchResultInterface {
+    public boolean onSearchQuery(String query);
+}
