+++ app/src/main/res/layout/layout_item_tag_result.xml
@@ -10,11 +10,13 @@
     <TextView android:id="@+id/tvTag"
         android:textSize="14sp"
         android:textColor="@color/gray_text"
+        android:fontFamily="sans-serif-medium"
         android:layout_width="match_parent"
         android:layout_height="wrap_content" />
 
     <TextView android:id="@+id/tvCount"
         android:textSize="14sp"
+        android:fontFamily="sans-serif"
         android:textColor="@color/light_gray_text"
         android:layout_width="match_parent"
         android:layout_height="wrap_content" />
