+++ app/src/main/res/layout/layout_item_user.xml
@@ -33,12 +33,14 @@
             <TextView android:id="@+id/tvUserNameSearchResult"
                 android:textSize="14sp"
                 android:textColor="@color/gray_text"
+                android:fontFamily="sans-serif-medium"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content" />
 
             <TextView android:id="@+id/tvFullName"
                 android:textSize="14sp"
                 android:textColor="@color/light_gray_text"
+                android:fontFamily="sans-serif"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content" />
